requirements.txt
# Core dependencies
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
python-dotenv==1.0.0
requests==2.32.3
Jinja2==3.1.4
Werkzeug==3.0.6
